Created a new package util/diff to carry Diff functionality
In the light of a potential spreading of Diff and MultiDiff approach to
some other controllers rather than application_controller, the generic
part of Diff'ing was moved to a stand-alone package.
